adjust the lenght of a "second" to account for clock drift.
I assume that the processing time to carry out all of the branching and computation of how to
display time, also takes some time. Because of this and that I rely on a call to "delay()" to keep
time, I need to shorten what the delay is, to account for processing time.